requirements.txt
# Core Framework & Server
Flask>=3.1.0
python-dotenv>=1.1.0
gunicorn>=23.0.0
Werkzeug>=3.1.0

# Database & ORM
Flask-SQLAlchemy>=3.1.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9.10
Flask-Migrate>=4.1.0
pytz>=2025.2

# Authentication
Flask-JWT-Extended>=4.7.0

# Validation & Serialization
marshmallow>=4.0.0

# AI / External APIs
google-generativeai>=0.8.5
google-genai>=1.12.1
requests>=2.31.0
protobuf~=5.29.4

# Utilities
flask-cors>=5.0.0